Move extension modules to extension namespace
