+++ FoliCon/Modules/Extension/FocusExtension.cs
@@ -1,7 +1,7 @@
 ﻿using NLog;
 using Logger = NLog.Logger;
 
-namespace FoliCon.Modules;
+namespace FoliCon.Modules.Extension;
 
 /// <summary>
 /// This class registers an attached property that can be set to trigger
