+++ FoliCon/ViewModels/CustomIconControlViewModel.cs
@@ -1,5 +1,6 @@
 ﻿// ReSharper disable SwitchStatementMissingSomeEnumCasesNoDefault
 
+using FoliCon.Modules.Extension;
 using FoliCon.Modules.utils;
 using GongSolutions.Wpf.DragDrop;
 using NLog;
