+++ FoliCon/Views/ApiConfiguration.xaml
@@ -7,7 +7,7 @@
              xmlns:mc="http://schemas.openxmlformats.org/markup-compatibility/2006"
              xmlns:viewModels="clr-namespace:FoliCon.ViewModels"
              xmlns:langs="clr-namespace:FoliCon.Properties.Langs"
-             xmlns:langExtension="clr-namespace:FoliCon.Modules.LangExtension"
+             xmlns:extension="clr-namespace:FoliCon.Modules.Extension"
              mc:Ignorable="d"
              prism:ViewModelLocator.AutoWireViewModel="True" HorizontalAlignment="Stretch" VerticalAlignment="Stretch"
              Background="{DynamicResource RegionBrush}"
@@ -33,10 +33,10 @@
             <StackPanel HorizontalAlignment="Stretch" VerticalAlignment="Stretch">
                 <hc:TextBox hc:InfoElement.Necessary="True"
                             Text="{Binding DArtClient, Mode=TwoWay,UpdateSourceTrigger=PropertyChanged}"
-                            Style="{StaticResource TextBoxExtend}" hc:InfoElement.Title="{langExtension:Lang Key={x:Static langs:LangKeys.ClientSecret}}" />
+                            Style="{StaticResource TextBoxExtend}" hc:InfoElement.Title="{extension:Lang Key={x:Static langs:LangKeys.ClientSecret}}" />
                 <hc:TextBox hc:InfoElement.Necessary="True"
                             Text="{Binding DArtClientId, Mode=TwoWay,UpdateSourceTrigger=PropertyChanged}"
-                            Style="{StaticResource TextBoxExtend}" hc:InfoElement.Title="{langExtension:Lang Key={x:Static langs:LangKeys.ClientID}}" />
+                            Style="{StaticResource TextBoxExtend}" hc:InfoElement.Title="{extension:Lang Key={x:Static langs:LangKeys.ClientID}}" />
             </StackPanel>
         </GroupBox>
         <GroupBox Style="{StaticResource GroupBoxTab}" Header="IGDB(Twitch) Api" Grid.Row="1"
@@ -45,19 +45,19 @@
             <StackPanel HorizontalAlignment="Stretch" VerticalAlignment="Stretch">
                 <hc:TextBox hc:InfoElement.Necessary="True"
                             Text="{Binding IgdbClientSecret, Mode=TwoWay,UpdateSourceTrigger=PropertyChanged}"
-                            Style="{StaticResource TextBoxExtend}" hc:InfoElement.Title="{langExtension:Lang Key={x:Static langs:LangKeys.ClientSecret}}" />
+                            Style="{StaticResource TextBoxExtend}" hc:InfoElement.Title="{extension:Lang Key={x:Static langs:LangKeys.ClientSecret}}" />
                 <hc:TextBox hc:InfoElement.Necessary="True"
                             Text="{Binding IgdbClientId, Mode=TwoWay,UpdateSourceTrigger=PropertyChanged}"
-                            Style="{StaticResource TextBoxExtend}" hc:InfoElement.Title="{langExtension:Lang Key={x:Static langs:LangKeys.ClientID}}" />
+                            Style="{StaticResource TextBoxExtend}" hc:InfoElement.Title="{extension:Lang Key={x:Static langs:LangKeys.ClientID}}" />
             </StackPanel>
         </GroupBox>
         <hc:TextBox hc:InfoElement.Necessary="True"
                     Text="{Binding TmdbKey,Mode=TwoWay,UpdateSourceTrigger=PropertyChanged}" Grid.Row="2"
-                    hc:InfoElement.Title="{langExtension:Lang Key={x:Static langs:LangKeys.TMDBAPIKey}}" Margin="8,7,8,0" />
-        <Button Grid.Row="3" Command="{Binding SaveCommand}" Content="{langExtension:Lang Key={x:Static langs:LangKeys.Save}}" Style="{StaticResource ButtonPrimary}"
+                    hc:InfoElement.Title="{extension:Lang Key={x:Static langs:LangKeys.TMDBAPIKey}}" Margin="8,7,8,0" />
+        <Button Grid.Row="3" Command="{Binding SaveCommand}" Content="{extension:Lang Key={x:Static langs:LangKeys.Save}}" Style="{StaticResource ButtonPrimary}"
                  Width="75" Margin="266,5,0,0" FontSize="16" HorizontalAlignment="Left" Height="auto"
                 VerticalAlignment="Top" IsDefault="True" TextOptions.TextRenderingMode="Aliased"/>
-        <Button Grid.Row="3" Content="{langExtension:Lang Key={x:Static langs:LangKeys.Close}}" Command="{Binding CloseDialogCommand}" CommandParameter="false"
+        <Button Grid.Row="3" Content="{extension:Lang Key={x:Static langs:LangKeys.Close}}" Command="{Binding CloseDialogCommand}" CommandParameter="false"
                 Style="{StaticResource ButtonDanger}" HorizontalAlignment="Left" Margin="372,5,0,0" Width="75" Height="auto"
                 FontSize="16" VerticalAlignment="Top" IsCancel="True" />
     </Grid>
